refactor: Separate ast nodes into multiple directories
