+++ src/core/ast/attribute.rs
@@ -1,0 +1,6 @@
+
+#[derive(PartialEq, Eq, Debug, Clone)]
+pub enum Attribute {
+    Flag(String),
+    KeyValue(String, String),
+}
