+++ src/core/ast/program.rs
@@ -1,0 +1,4 @@
+use crate::core::{ast::statement::Statement, utils::span::Span};
+
+#[derive(PartialEq, Eq, Debug)]
+pub struct Program(pub Vec<Statement>, pub Span);
