+++ src/core/constraints/constraint.rs
@@ -4,7 +4,7 @@ use std::{collections::HashMap, fmt::Debug};
 
 use crate::{
     core::{
-        ast::nodes::{ConstraintKind, DataTypeKind},
+        ast::{ConstraintKind, DataTypeKind},
         constraints::sampler::{BooleanSampler, IdentitySampler, Sampler, UniformSampler},
         utils::extract_object,
     },
